Update openCypher-Exercises-Answer-Key.ipynb
As mentioned earlier in the notebook, the `MERGE` clause does not support partial pattern matches.  When matching against patterns in `MERGE`, the matches are either the entire pattern, or the entire pattern is created.

Because the pattern is a partial match, it currently creates a new "person {first_name: 'Dave'}" node for each of the new "person" added to the graph. 

By first doing a MERGE for (d:person {first_name: 'Dave'), when the partial match happens, the whole path is again created, but this time linking to the already existing "Dave" rather than creating new ones each time.